+++ packages/bluesky/lib/src/entities/profile_record.freezed.dart
@@ -35,6 +35,10 @@ mixin _$ProfileRecord {
   /// Represents the avatar image of the actor.
   Blob? get avatar => throw _privateConstructorUsedError;
 
+  /// Attached labels.
+  @labelsConverter
+  Labels? get labels => throw _privateConstructorUsedError;
+
   /// Represents the banner image of the actor.
   Blob? get banner => throw _privateConstructorUsedError;
 
@@ -55,9 +59,11 @@ abstract class $ProfileRecordCopyWith<$Res> {
       String? displayName,
       String? description,
       Blob? avatar,
+      @labelsConverter Labels? labels,
       Blob? banner});
 
   $BlobCopyWith<$Res>? get avatar;
+  $LabelsCopyWith<$Res>? get labels;
   $BlobCopyWith<$Res>? get banner;
 }
 
@@ -78,6 +84,7 @@ class _$ProfileRecordCopyWithImpl<$Res, $Val extends ProfileRecord>
     Object? displayName = freezed,
     Object? description = freezed,
     Object? avatar = freezed,
+    Object? labels = freezed,
     Object? banner = freezed,
   }) {
     return _then(_value.copyWith(
@@ -97,6 +104,10 @@ class _$ProfileRecordCopyWithImpl<$Res, $Val extends ProfileRecord>
           ? _value.avatar
           : avatar // ignore: cast_nullable_to_non_nullable
               as Blob?,
+      labels: freezed == labels
+          ? _value.labels
+          : labels // ignore: cast_nullable_to_non_nullable
+              as Labels?,
       banner: freezed == banner
           ? _value.banner
           : banner // ignore: cast_nullable_to_non_nullable
@@ -113,6 +124,18 @@ class _$ProfileRecordCopyWithImpl<$Res, $Val extends ProfileRecord>
 
     return $BlobCopyWith<$Res>(_value.avatar!, (value) {
       return _then(_value.copyWith(avatar: value) as $Val);
+    });
+  }
+
+  @override
+  @pragma('vm:prefer-inline')
+  $LabelsCopyWith<$Res>? get labels {
+    if (_value.labels == null) {
+      return null;
+    }
+
+    return $LabelsCopyWith<$Res>(_value.labels!, (value) {
+      return _then(_value.copyWith(labels: value) as $Val);
     });
   }
 
@@ -142,10 +165,13 @@ abstract class _$$_ProfileRecordCopyWith<$Res>
       String? displayName,
       String? description,
       Blob? avatar,
+      @labelsConverter Labels? labels,
       Blob? banner});
 
   @override
   $BlobCopyWith<$Res>? get avatar;
+  @override
+  $LabelsCopyWith<$Res>? get labels;
   @override
   $BlobCopyWith<$Res>? get banner;
 }
@@ -165,6 +191,7 @@ class __$$_ProfileRecordCopyWithImpl<$Res>
     Object? displayName = freezed,
     Object? description = freezed,
     Object? avatar = freezed,
+    Object? labels = freezed,
     Object? banner = freezed,
   }) {
     return _then(_$_ProfileRecord(
@@ -184,6 +211,10 @@ class __$$_ProfileRecordCopyWithImpl<$Res>
           ? _value.avatar
           : avatar // ignore: cast_nullable_to_non_nullable
               as Blob?,
+      labels: freezed == labels
+          ? _value.labels
+          : labels // ignore: cast_nullable_to_non_nullable
+              as Labels?,
       banner: freezed == banner
           ? _value.banner
           : banner // ignore: cast_nullable_to_non_nullable
@@ -201,6 +232,7 @@ class _$_ProfileRecord implements _ProfileRecord {
       this.displayName,
       this.description,
       this.avatar,
+      @labelsConverter this.labels,
       this.banner});
 
   factory _$_ProfileRecord.fromJson(Map<String, dynamic> json) =>
@@ -225,13 +257,18 @@ class _$_ProfileRecord implements _ProfileRecord {
   @override
   final Blob? avatar;
 
+  /// Attached labels.
+  @override
+  @labelsConverter
+  final Labels? labels;
+
   /// Represents the banner image of the actor.
   @override
   final Blob? banner;
 
   @override
   String toString() {
-    return 'ProfileRecord(type: $type, displayName: $displayName, description: $description, avatar: $avatar, banner: $banner)';
+    return 'ProfileRecord(type: $type, displayName: $displayName, description: $description, avatar: $avatar, labels: $labels, banner: $banner)';
   }
 
   @override
@@ -245,13 +282,14 @@ class _$_ProfileRecord implements _ProfileRecord {
             (identical(other.description, description) ||
                 other.description == description) &&
             (identical(other.avatar, avatar) || other.avatar == avatar) &&
+            (identical(other.labels, labels) || other.labels == labels) &&
             (identical(other.banner, banner) || other.banner == banner));
   }
 
   @JsonKey(ignore: true)
   @override
-  int get hashCode =>
-      Object.hash(runtimeType, type, displayName, description, avatar, banner);
+  int get hashCode => Object.hash(
+      runtimeType, type, displayName, description, avatar, labels, banner);
 
   @JsonKey(ignore: true)
   @override
@@ -273,6 +311,7 @@ abstract class _ProfileRecord implements ProfileRecord {
       final String? displayName,
       final String? description,
       final Blob? avatar,
+      @labelsConverter final Labels? labels,
       final Blob? banner}) = _$_ProfileRecord;
 
   factory _ProfileRecord.fromJson(Map<String, dynamic> json) =
@@ -297,6 +336,11 @@ abstract class _ProfileRecord implements ProfileRecord {
 
   /// Represents the avatar image of the actor.
   Blob? get avatar;
+  @override
+
+  /// Attached labels.
+  @labelsConverter
+  Labels? get labels;
   @override
 
   /// Represents the banner image of the actor.