+++ packages/bluesky/lib/src/params/thread_param.g.dart
@@ -26,6 +26,10 @@ _$_ThreadParam _$$_ThreadParamFromJson(Map json) => $checkedCreate(
                   v, embedConverter.fromJson)),
           languageTags: $checkedConvert('languageTags',
               (v) => (v as List<dynamic>?)?.map((e) => e as String).toList()),
+          labels: $checkedConvert(
+              'labels',
+              (v) => _$JsonConverterFromJson<Map<String, dynamic>, Labels>(
+                  v, labelsConverter.fromJson)),
           createdAt: $checkedConvert('createdAt',
               (v) => v == null ? null : DateTime.parse(v as String)),
           unspecced: $checkedConvert(
@@ -57,6 +61,10 @@ Map<String, dynamic> _$$_ThreadParamToJson(_$_ThreadParam instance) {
       _$JsonConverterToJson<Map<String, dynamic>, Embed>(
           instance.embed, embedConverter.toJson));
   writeNotNull('languageTags', instance.languageTags);
+  writeNotNull(
+      'labels',
+      _$JsonConverterToJson<Map<String, dynamic>, Labels>(
+          instance.labels, labelsConverter.toJson));
   writeNotNull('createdAt', instance.createdAt?.toIso8601String());
   val['unspecced'] = instance.unspecced;
   return val;